+++ background/city_map.js
@@ -7,7 +7,10 @@ const cityMap = {
     }
     switch (method) {
       case 'updateEntity':
-        greatBuilding.checkArcBonus(data)
+        if (startup.playerId === data[0].player_id) {
+          // only update our own info
+          greatBuilding.checkArcBonus(data)
+        }
         greatBuilding.storeBuildingInfo(id, data[0].player_id, data[0].state.forge_points_for_level_up)
         break
       case 'reset':