+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Various other/minor GB information improvements
 * Check for new version
 * Update GB and Boost information when spending SP
+* Only update Arc bonus when updating owned Arc
 
 # 0.2.0
 * Rewards screen - Order on goods in stock when value is the same ([#2](https://github.com/veger/foei/issues/2))